GuardCheckLock Changes
TryLockTarget now looks through the radars, determines if they can lock the target, and then asks the radar to see if it has enough locks available, or if it can free up a lock. This minimizes the unlocks to only those necessary for an engagement and also ensures that a lock is actually available on a radar capable of locking the target. This supplants the previous role of GuardCheckLock, which now only performs a linear search to determine if the target is already locked.